reorganized files, added tests
