+++ packages/javascript/bh-shared-ui/src/hooks/useExploreGraph/index.ts
@@ -14,6 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './queries';
 export * from './useExploreGraph';
-
-export * from './search-modes';
+export * from './useNodeSearch';
