+++ packages/javascript/bh-shared-ui/src/hooks/useExploreGraph/queries/index.ts
@@ -14,6 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './node-search';
-export * from './useNodeSearch';
+export * from '../useNodeSearch/useNodeSearch';
+export * from './nodeSearch';
 export * from './utils';
